requirements.txt
appnope==0.1.2
autopep8==1.6.0
backcall==0.2.0
blpapi==3.18.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.12
decorator==5.0.5
et-xmlfile==1.0.1
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.53.0
greenlet==1.0.0
httplib2==0.19.1
ibapi==9.81.1.post1
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
lxml==4.6.3
multitasking==0.0.9
numpy==1.20.2
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.15.8
psycopg2-binary==2.8.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.6
toml==0.10.2
tqdm==4.60.0
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
XlsxWriter==1.4.3
yfinance==0.1.70
